requirements.txt
streamlit
google-generativeai
python-dotenv
duckduckgo-search
PyGithub
